requirements.txt
fastapi
uvicorn[standard]
httpx
pyyaml
psycopg2-binary
pytest
pytest-asyncio
